cmd/sunlight: optionally validate private and public keys match (#6)
I expect that the configuration file will be in config management systems, but
the private key will be more controlled. Because we often need to know the
public key, it's helpful to have configuration that's self-documenting and
ensures there's no mixups in what keys get used.

The new publickey configuration option is optional.