internal/ctlog: drop overwrite protection on Tigris
This can cause the log to lock up as describe in #12, or even in steady
state due to request hedging: if request A wins, but request B returns a
412 before request A's 200, the sequencing will fail (safely) and leave
behind non-overwritable tiles.

This means the tree might suffer fatal failures on Tigris if two
sequencers run at the same time, or if the stale Get issues in the
LockBackend reoccur, since there is no object versioning.