internal/ctlog: raise sequencer timeout
This is arguably not the right long-term solution, because 1) we should
probably only raise the timeout for the final tiles upload, where
failure is fatal, and lower it if anything for other uploads, and 2) a
longer sequencing will make the next pool bigger, which might be bad at
a time of increased backend latency. However, for now it should let us
observe the latency spikes better.